Toggles the backlight a given number of times
Uses the report payload to determine how many times to toggle the
backlight. Note that during the backlight toggling, the keyboard is
incapable of sending any keystrokes. This is likely either because
we are blocking some important thread while we do wait_ms or because
of some issues we've caused with console_flush atomicity.